make strict_path_check rule specific, fix bug in option handling
Makes strict_path_check specific to an individual audit rule, instead of
a global option for the rule set.

In SetRules, return an additional argument indicating soft errors, such
as when a rule has been skipped.

This also fixes the way the option was being handled, previously if
strict_path_check was enabled and the watch rule could not find the
file, it would continue instead of returning resulting in a panic.

Closes #18