handle cases of partial read from kernel netlink producer
Closes #24

In some cases, the netlink payload value in a message received from the
kernel is not what it is expected to be.

This seems to be dependent on where the audit message is being created
in the kernel. For example, some functions make use of netlink macros
which correctly add the header to the payload size (inclusive of
header), and others such as audit_log_end appear to manually adjust the
netlink message size, and set it to the size of the skb.

This patch adjusts the behavior of the message consumer to simply use
all bytes in the payload in the received buffer and ignore the netlink
message length, unless the message is NLM_F_MULTI (multipart).